Add files via upload
Overview — Many of us are dealing with instructional redesign as we transition from what was the traditional classroom. We often measure the effectiveness of the instruction through some form of diagnostic test. The type of test we use is dependent upon how best to measure the desired outcome. This test might be, for example, by observation and a checklist, or perhaps something like the familiar multiple choice test. The results of this testing are typically some form of right/wrong correctness scoring. The intent of this diagnostic software, CatTrax, is to use the same testing materials, yet surpass simple test correctness scores to help identify not only what is easy or hard, but areas of curricular confusion, and learners who may be experiencing confusion. This means that potentially “at-risk” learners can be identified at an early stage when intervention is most effective. Used in a formative manner, new learning can be better tailored to diverse settings and learner-centered methods.

How it works — Your data file, formatted to a common CSV spreadsheet format, contains participant identification, test results, and a single or multiple answer keys. The help file provides explanation for creating a data file for various test types, though the data file could certainly be formatted for other purposes. 

The analysis of this data is then done on-the-fly to create a browser-based interpretation with the option of downloading all of the results as a CSV file for your further analysis. The results provide a straightforward verbal interpretation of the testing outcome, graphic tables, as well as traditional test measures. 

In both graphical and numeric formats, each participant shows their correctness score along with their Standard Error of Measurement and participant's true ability within a range of scores, their z-score, and their Caution Index. The Caution Index, established by Sato, Harnish and Linn, measures the consistency of the participant's responses giving insight into guessing, confusion, unusual instructional or experiential history, copying, high anxiety, or carelessness — used formatively, these inconsistencies are potentially useful for identifying “at-risk” participants at an early stage. 

In both graphical and numeric formats, each test item shows, the total number of participants responding correctly, Difficulty level, Discrimination Index, and Caution Index. Additionally, Item Analysis is also provided as split-half for groups of less than thirty people and split thirds for larger groups. The help file provides a much more detailed explanation for creating your data file and interpreting your results. 

Download the application — CatTrax diagnostics is a PHP server-based program. For minimalists, a newer Raspberry Pi 3 or 4 single board computer will run it with Apache or a lighter server. With the exception of the sample test data file (input.csv), the unzipped CatTrax contents reside in the server HTML directory. Diagnostic results are generated on the fly and viewable in your browser. Alternatively, CatTrax can run as a "non-server" application using PHPs built-in server.